add analytics, fix bug, display tip on top of scroll view
