+++ Nos/Views/Discover/DiscoverTab.swift
@@ -28,18 +28,26 @@ struct DiscoverTab: View {
     var body: some View {
         NosNavigationStack(path: $router.discoverPath) {
             ZStack {
-                VStack {
-                    DiscoverContentsView(
-                        featuredAuthorCategory: .all,
-                        searchController: searchController
-                    )
+                DiscoverContentsView(
+                    featuredAuthorCategory: .all,
+                    searchController: searchController
+                )
 
+                VStack {
+                    Spacer()
+                    
                     TipView(goToFeedTip)
                         .padding(.horizontal, 12)
                         .padding(.bottom, 8)
                         .readabilityPadding()
                         .tipBackground(LinearGradient.horizontalAccentReversed)
                         .tipViewStyle(.pointDownEmoji)
+                        .onAppear {
+                            analytics.displayedGoToFeedTip()
+                        }
+                        .onDisappear {
+                            analytics.dismissedGoToFeedTip()
+                        }
                 }
             }
             .searchable(
@@ -70,31 +78,6 @@ struct DiscoverTab: View {
             .toolbarBackground(.visible, for: .navigationBar)
             .toolbarBackground(Color.cardBgBottom, for: .navigationBar)
             .navigationBarItems(leading: SideMenuButton())
-        }
-    }
-}
-
-/// A tip that's displayed on the Discover view after the user has followed three accounts.
-struct GoToFeedTip: Tip {
-    /// A TipKit Event that tracks the number of accounts that have been followed.
-    static let followedAccount = Tips.Event(id: "followedAccount")
-
-    /// A TipKit Event that tracks how many times the Feed has been displayed.
-    static let viewedFeed = Tips.Event(id: "viewedFeed")
-
-    var title: Text {
-        Text("goToYourFeed")
-    }
-
-    var rules: [Rule] {
-        // Each rule here is combined using the logical AND, so all rules must return true for the tip to display.
-
-        #Rule(Self.followedAccount) {
-            $0.donations.count >= 3
-        }
-
-        #Rule(Self.viewedFeed) {
-            $0.donations.count < 2 // for whatever reason, HomeFeedView's onAppear is called before the user views it
         }
     }
 }
