+++ Nos/Views/Home/HomeFeedView.swift
@@ -74,6 +74,12 @@ struct HomeFeedView: View {
                     .readabilityPadding()
                     .tipBackground(LinearGradient.horizontalAccentReversed)
                     .tipViewStyle(.inline)
+                    .onAppear {
+                        analytics.displayedWelcomeToFeedTip()
+                    }
+                    .onDisappear {
+                        analytics.dismissedWelcomeToFeedTip()
+                    }
 
                 PagedNoteListView(
                     refreshController: $refreshController,
